avance ok

+++ src/main/java/Senet/Ramboot/entity/UsuarioEntity.java
@@ -10,6 +10,7 @@ import org.antlr.v4.runtime.misc.NotNull;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -77,7 +78,7 @@ public class UsuarioEntity {
     private boolean isActive;
 
 
-    private double saldo;
+    private BigDecimal saldo;
 
     @OneToMany(mappedBy = "usuario", fetch = FetchType.LAZY)
     private List<GcontrataEntity> Gcontrata = new ArrayList<>();
@@ -89,7 +90,7 @@ public class UsuarioEntity {
 
     public UsuarioEntity() {}
 
-    public UsuarioEntity(String username, String nombre, String apellido1, String apellido2, String email, boolean isActive, String DNI, String feedback, String CP, double saldo, Long telefono, TipousuarioEntity tipousuario) {
+    public UsuarioEntity(String username, String nombre, String apellido1, String apellido2, String email, boolean isActive, String DNI, String feedback, String CP, BigDecimal saldo, Long telefono, TipousuarioEntity tipousuario) {
         this.username = username;   
         this.nombre = nombre;
         this.apellido1 = apellido1;
@@ -106,7 +107,7 @@ public class UsuarioEntity {
         this.ultima_conexion = null;
     }
 
-    public UsuarioEntity(Long id, String username, String nombre, String apellido1, String apellido2, String email, String DNI, String feedback, String CP,double saldo, Long telefono, TipousuarioEntity tipousuario) {
+    public UsuarioEntity(Long id, String username, String nombre, String apellido1, String apellido2, String email, String DNI, String feedback, String CP,BigDecimal saldo, Long telefono, TipousuarioEntity tipousuario) {
         this.id = id;
         this.username = username;   
         this.nombre = nombre;
@@ -186,11 +187,11 @@ public class UsuarioEntity {
         this.isActive = isActive;
     }
 
-    public double getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(double saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
